Only try and compute a date if the value is numeric.
Otherwise we'll make a broken date like the instant at the start of the epoch for values like '' or null.

This is less forgiving of tokens which have a number-shaped string for the timestamp. Since the JWT spec says "a JSON numeric value" it's probably better to call that invalid than to coerce the type.